+++ components/footer.vue
@@ -2,7 +2,8 @@
 	<footer class="footer">
 		<div class="content has-text-centered">
 			<p>
-				Copyright &COPY; {{ new Date().getFullYear() }} <strong>Huawei Open Day: Intelligent World 2030</strong>. All rights reserved.
+				Copyright &COPY; {{ new Date().getFullYear() }} <strong>Huawei Open Day: Intelligent World 2030</strong>. All
+				rights reserved.
 			</p>
 		</div>
 	</footer>
