+++ package.json
@@ -12,7 +12,7 @@
 		"start": "pnpm run css-watch"
 	},
 	"devDependencies": {
-		"@types/node": "^18.16.6",
+		"@types/node": "^18",
 		"nuxt": "^3.4.3"
 	},
 	"dependencies": {
